% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeepOperators.R
\docType{package}
\name{deepoperators}
\alias{deepoperators}
\alias{deepoperators-package}
\title{DeepOperators: Deeply Learned Logical Operators

These operators act on raw, logical and number-like vectors.}
\description{
DeepOperators: Deeply Learned Logical Operators

These operators act on raw, logical and number-like vectors.
}
